Refactored log widget and added to main charting page
